refactor: RenderPipeline
