+++ crates/core/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod accessibility;
-pub mod compositor;
 pub mod dom;
 pub mod elements;
 pub mod events;
@@ -8,14 +7,12 @@ pub mod node;
 pub mod platform_state;
 pub mod plugins;
 pub mod render;
-pub mod skia;
 pub mod style;
 pub mod types;
 
 pub mod prelude {
     pub use crate::{
         accessibility::*,
-        compositor::*,
         dom::*,
         elements::*,
         events::*,
@@ -24,7 +21,6 @@ pub mod prelude {
         platform_state::*,
         plugins::*,
         render::*,
-        skia::*,
         style::*,
         types::*,
     };
