+++ crates/core/src/render/compositor.rs
