+++ crates/core/src/render/wireframe_renderer.rs
