--- crates/core/src/skia/mod.rs
@@ -1,8 +1,0 @@
-pub mod paragraph;
-pub mod skia_measurer;
-pub mod skia_renderer;
-mod wireframe_renderer;
-
-pub use paragraph::*;
-pub use skia_measurer::*;
-pub use skia_renderer::*;
